requirements.txt
# Core Django
Django==4.2.7
python-decouple==3.8
Pillow==10.1.0

# Face Recognition
deepface==0.0.95
tf-keras==2.17.0
tensorflow==2.17.0

# Image Processing
opencv-python-headless==4.8.1.78
numpy==1.24.3

# Cloud Storage
cloudinary==1.41.0

# HTTP Requests
requests==2.31.0

# Forms and Validation
django-crispy-forms==2.1

# File Handling
django-cleanup==8.0.0

# Security
django-cors-headers==4.3.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0

